Implement live response generation for eval system
- Replace stored responses with fresh generation via eval endpoint
- Update test format to only store context and requirements
- Clean up old format files and temporary scripts
- Ensure evals test current CJ behavior, not historical responses
