--- hirecj_evals/datasets/all_tests_new.jsonl
@@ -1,2 +1,0 @@
-{"sample_id": "greeting_test", "context": {"messages": [{"role": "user", "content": "sup guy"}], "workflow": "ad_hoc_support", "persona": "jessica"}, "requirements": ["Must not say 'As CJ I'd say' or similar meta-commentary", "Must greet the user in a friendly and professional manner", "Must identify as CJ"]}
-{"sample_id": "day_question_test", "context": {"messages": [{"role": "user", "content": "sup guy"}, {"role": "assistant", "content": "Hey there! CJ here – ready to jump in and help you tackle whatever comes your way. What's on your mind today?"}, {"role": "user", "content": "what day is it"}], "workflow": "ad_hoc_support", "persona": "jessica"}, "requirements": ["Must not say 'As CJ I'd say' or similar meta-commentary", "Must provide the day of the week or indicate inability to know current date", "Must offer to help with something else"]}
