+++ scripts/add_test.py
@@ -1,15 +1,15 @@
 #!/usr/bin/env python3
 """
 Interactive script to add test cases with plain English requirements.
+Only saves conversation context, not CJ's responses.
 """
 
 import json
 import sys
 from pathlib import Path
 from datetime import datetime
-from typing import List, Dict, Any
+from typing import Dict, List, Any, Tuple, Optional
 
-# Add parent directory to path
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
 from rich.console import Console
@@ -19,147 +19,116 @@ from rich.prompt import Prompt, Confirm
 console = Console()
 
 # Paths
-CONVERSATIONS_DIR = Path('hirecj_evals/conversations/playground')
+CONVERSATIONS_BASE = Path('hirecj_evals/conversations')
 ALL_TESTS_FILE = Path('hirecj_evals/datasets/all_tests.jsonl')
 
 
-def find_recent_conversations(limit: int = 10) -> List[Path]:
+def find_recent_conversations(limit: int = 10) -> List[Tuple[Path, Dict[str, Any]]]:
     """Find recent conversation files."""
     conversations = []
     
-    # Get all JSON files from all date directories
-    for date_dir in CONVERSATIONS_DIR.glob('*'):
-        if date_dir.is_dir():
+    # Search through all source directories
+    for source_dir in CONVERSATIONS_BASE.glob('*'):
+        if not source_dir.is_dir():
+            continue
+            
+        # Search through date directories
+        for date_dir in source_dir.glob('*'):
+            if not date_dir.is_dir():
+                continue
+                
+            # Find JSON files
             for conv_file in date_dir.glob('*.json'):
-                if not conv_file.name.startswith('test_'):  # Skip test files
-                    conversations.append(conv_file)
+                try:
+                    with open(conv_file, 'r') as f:
+                        data = json.load(f)
+                    
+                    # Extract basic info
+                    conv_info = {
+                        'file': conv_file,
+                        'timestamp': data.get('timestamp', ''),
+                        'message_count': len(data.get('messages', [])),
+                        'workflow': data.get('context', {}).get('workflow', {}).get('name', 'unknown')
+                    }
+                    conversations.append((conv_file, conv_info))
+                except Exception as e:
+                    console.print(f"[red]Error reading {conv_file}: {e}[/red]")
     
-    # Sort by modification time, most recent first
-    conversations.sort(key=lambda x: x.stat().st_mtime, reverse=True)
-    
+    # Sort by timestamp descending and return top N
+    conversations.sort(key=lambda x: x[1]['timestamp'], reverse=True)
     return conversations[:limit]
 
 
-def display_conversation(conv_data: Dict[str, Any]) -> None:
-    """Display a conversation in a readable format."""
-    messages = conv_data.get('messages', [])
+def extract_conversation_context(conv_data: Dict[str, Any]) -> List[Dict[str, Any]]:
+    """Extract context (user messages only) from conversation data."""
+    context_messages = []
     
-    console.print(f"\n[bold cyan]Conversation ID:[/bold cyan] {conv_data.get('id', 'Unknown')}")
-    console.print(f"[bold cyan]Timestamp:[/bold cyan] {conv_data.get('timestamp', 'Unknown')}")
-    console.print(f"[bold cyan]Total turns:[/bold cyan] {len(messages)}\n")
-    
-    for i, msg in enumerate(messages):
-        console.print(f"[bold yellow]Turn {msg['turn']}:[/bold yellow]")
-        console.print(f"[green]User:[/green] {msg['user_input']}")
+    for msg in conv_data.get('messages', []):
+        # Add user message
+        context_messages.append({
+            "role": "user",
+            "content": msg['user_input']
+        })
         
-        response = msg['agent_processing']['final_response']
-        # Truncate long responses
-        if len(response) > 200:
-            response = response[:197] + "..."
-        console.print(f"[blue]CJ:[/blue] {response}\n")
+        # Add assistant message (but not the agent_processing details)
+        if msg.get('agent_processing', {}).get('final_response'):
+            context_messages.append({
+                "role": "assistant", 
+                "content": msg['agent_processing']['final_response']
+            })
+    
+    return context_messages
 
 
-def select_conversation() -> Dict[str, Any]:
-    """Let user select a conversation."""
+def select_conversation() -> Tuple[Dict[str, Any], Path]:
+    """Let user select a conversation to convert."""
     conversations = find_recent_conversations()
     
     if not conversations:
         console.print("[red]No conversations found![/red]")
         sys.exit(1)
     
-    # Display table of conversations
+    # Display table
     table = Table(title="Recent Conversations")
-    table.add_column("", style="cyan", width=3)
+    table.add_column("#", style="cyan", width=5)
     table.add_column("File", style="green")
     table.add_column("Date", style="yellow")
     table.add_column("Time", style="yellow")
     
-    for i, conv_path in enumerate(conversations):
-        mod_time = datetime.fromtimestamp(conv_path.stat().st_mtime)
+    for i, (path, info) in enumerate(conversations, 1):
+        timestamp = info['timestamp']
+        if timestamp:
+            try:
+                dt = datetime.fromisoformat(timestamp.replace('Z', '+00:00'))
+                date = dt.strftime('%Y-%m-%d')
+                time = dt.strftime('%H:%M:%S')
+            except:
+                date = time = "Unknown"
+        else:
+            date = time = "Unknown"
+            
         table.add_row(
-            str(i + 1),
-            conv_path.stem,
-            mod_time.strftime('%Y-%m-%d'),
-            mod_time.strftime('%H:%M:%S')
+            str(i),
+            path.stem,
+            date,
+            time
         )
     
     console.print(table)
     
     # Get selection
-    while True:
-        choice = Prompt.ask("\nSelect conversation", default="1")
-        try:
-            idx = int(choice) - 1
-            if 0 <= idx < len(conversations):
-                break
-            console.print("[red]Invalid selection![/red]")
-        except ValueError:
-            console.print("[red]Please enter a number![/red]")
+    choice = Prompt.ask("\nSelect conversation", default="1")
+    try:
+        idx = int(choice) - 1
+        if 0 <= idx < len(conversations):
+            path, _ = conversations[idx]
+            with open(path, 'r') as f:
+                return json.load(f), path
+    except:
+        pass
     
-    # Load and return conversation
-    with open(conversations[idx], 'r') as f:
-        return json.load(f), conversations[idx]
-
-
-def get_requirements() -> List[str]:
-    """Get requirements from user."""
-    requirements = []
-    
-    console.print("\n[bold cyan]Enter requirements in plain English[/bold cyan]")
-    console.print("[dim]Press Enter with empty line to finish[/dim]\n")
-    
-    while True:
-        req = Prompt.ask(f"Requirement {len(requirements) + 1}", default="")
-        if not req:
-            if requirements:
-                break
-            console.print("[yellow]Please enter at least one requirement![/yellow]")
-        else:
-            requirements.append(req)
-            console.print(f"[green]✓ Added:[/green] {req}")
-    
-    return requirements
-
-
-def create_test_cases(conv_data: Dict[str, Any], conv_path: Path, requirements: List[str]) -> List[Dict[str, Any]]:
-    """Create test cases from conversation with requirements."""
-    test_cases = []
-    messages = conv_data.get('messages', [])
-    
-    # Extract base filename
-    base_name = conv_path.stem
-    
-    # Build conversation history
-    system_prompt = conv_data['prompts']['cj_prompt'] + '\n\n' + conv_data['prompts']['workflow_prompt']
-    conversation_history = [{"role": "system", "content": system_prompt}]
-    
-    for msg in messages:
-        # Add user message
-        conversation_history.append({
-            "role": "user",
-            "content": msg['user_input']
-        })
-        
-        # Create test case for this turn
-        test_case = {
-            "sample_id": f"{base_name}_turn_{msg['turn']}",
-            "input": {
-                "messages": conversation_history.copy()
-            },
-            "actual": {
-                "response": msg['agent_processing']['final_response']
-            },
-            "requirements": requirements  # Same requirements for all turns
-        }
-        test_cases.append(test_case)
-        
-        # Add assistant response for next turn
-        conversation_history.append({
-            "role": "assistant",
-            "content": msg['agent_processing']['final_response']
-        })
-    
-    return test_cases
+    console.print("[red]Invalid selection![/red]")
+    sys.exit(1)
 
 
 def main():
@@ -168,41 +137,71 @@ def main():
     
     # Select conversation
     conv_data, conv_path = select_conversation()
-    display_conversation(conv_data)
+    console.print(f"\n[green]Selected:[/green] {conv_path.name}")
     
-    # Get requirements
-    requirements = get_requirements()
+    # Extract context
+    context_messages = extract_conversation_context(conv_data)
     
-    # Show summary
-    console.print("\n[bold cyan]Summary:[/bold cyan]")
-    console.print(f"Conversation: {conv_path.stem}")
-    console.print(f"Turns: {len(conv_data.get('messages', []))}")
-    console.print("Requirements:")
-    for req in requirements:
-        console.print(f"  • {req}")
+    # Show messages
+    console.print("\n[bold]Conversation Context:[/bold]")
+    for i, msg in enumerate(context_messages):
+        role_color = "blue" if msg["role"] == "user" else "green"
+        console.print(f"[{role_color}]{msg['role']}:[/{role_color}] {msg['content'][:100]}...")
     
-    # Confirm
-    if not Confirm.ask("\nAdd these test cases?"):
-        console.print("[yellow]Cancelled[/yellow]")
-        return
+    # Get workflow info
+    workflow = conv_data.get('context', {}).get('workflow', {}).get('name', 'ad_hoc_support')
+    persona = conv_data.get('context', {}).get('persona', {}).get('name', 'jessica')
     
-    # Create test cases
-    test_cases = create_test_cases(conv_data, conv_path, requirements)
+    # Collect requirements for each turn
+    test_cases = []
     
-    # Ensure directory exists
-    ALL_TESTS_FILE.parent.mkdir(parents=True, exist_ok=True)
+    # Process each user message
+    user_turns = [(i, msg) for i, msg in enumerate(context_messages) if msg["role"] == "user"]
     
-    # Append to all_tests.jsonl
-    with open(ALL_TESTS_FILE, 'a') as f:
-        for case in test_cases:
-            f.write(json.dumps(case) + '\n')
+    for turn_num, (msg_idx, user_msg) in enumerate(user_turns, 1):
+        console.print(f"\n[bold yellow]Turn {turn_num}:[/bold yellow]")
+        console.print(f"User: {user_msg['content']}")
+        
+        # Get messages up to this point (context for this turn)
+        turn_context = context_messages[:msg_idx + 1]
+        
+        # Ask for requirements
+        console.print("\n[cyan]Enter requirements for CJ's response (one per line, empty line to finish):[/cyan]")
+        requirements = []
+        while True:
+            req = Prompt.ask("Requirement", default="")
+            if not req:
+                break
+            requirements.append(req)
+        
+        if requirements:
+            test_case = {
+                "sample_id": f"{conv_path.stem}_turn_{turn_num}",
+                "context": {
+                    "messages": turn_context,
+                    "workflow": workflow,
+                    "persona": persona
+                },
+                "requirements": requirements
+            }
+            test_cases.append(test_case)
     
-    console.print(f"\n[green]✓ Added {len(test_cases)} test cases to {ALL_TESTS_FILE}[/green]")
-    
-    # Show what will be tested
-    console.print("\n[bold]Test cases added:[/bold]")
-    for i, case in enumerate(test_cases):
-        console.print(f"  Turn {i+1}: Will check {len(requirements)} requirements")
+    # Save test cases
+    if test_cases:
+        console.print(f"\n[green]Adding {len(test_cases)} test cases to {ALL_TESTS_FILE}[/green]")
+        
+        # Ensure directory exists
+        ALL_TESTS_FILE.parent.mkdir(parents=True, exist_ok=True)
+        
+        # Append to file
+        with open(ALL_TESTS_FILE, 'a') as f:
+            for tc in test_cases:
+                f.write(json.dumps(tc) + '\n')
+        
+        console.print("[bold green]✓ Test cases added successfully![/bold green]")
+        console.print("\nRun 'make test-reqs' to evaluate these requirements")
+    else:
+        console.print("[yellow]No test cases created[/yellow]")
 
 
 if __name__ == '__main__':
